fix: timeout data type

+++ sentry-core/src/main/java/io/sentry/NoOpHub.java
@@ -59,7 +59,7 @@ class NoOpHub implements IHub {
   public void bindClient(SentryClient client) {}
 
   @Override
-  public void flush(int timeoutMills) {}
+  public void flush(long timeoutMills) {}
 
   @Override
   public IHub clone() {
